move manage subscription function
